Increment generation, avoid sorting fields, and general fixups
Sorting components isn't needed as long as their alignment
requirements falls bellow 64-bytes (practically everything) thus it's
better to keep the order in which they're defined such that component
order reflects locality while pushing 0-bit components to the end.

Keeping a generation count for entities prevents equality between
entity ids reissued near retirement and increases the chances for
assertions to trip.

Registration of buckets should be easier and as the model owns the
bucket there's no point in having the user pass in buckets for
use. However, buckets are still not allocated on demand such that it's
possible to decide if the new would-be allocation is a bug in the game
or not.