Implement various response curves This provides response curves to be used in the full Infinite Axis Utility System with several of the other parts missing such as:
- weighted action selection (part of dual utility)
- behaviour category selection (part of dual utility)
- application of noise to response curves (uniqueness)
- selecting multiple actions based on the resources available

Response curves define how valuable it is for a character to act given
a consideration. For example:

    EU = health / 100

results in a stronger desire the higher the health of the character
which may be part of a consideration of then to attack while:

    EU = 1 - health / 10

results in the opposite thus might be used to decide when to heal.
However, the above cannot be used to define a character that will seek
cover then low on health but go berserk when critical. For such, a
response curve such as the following is required to change the the
mapping from linear health to how the character should respond:

    x = (1 - health / 100) - 0.7
    EU = 4.5 * pow(x, 2) + -6.4 * pow(x, 3) + 3 * pow(x, 4) + 0.4

where the Expected Utility is high when health is near full and
when the health is below critical triggering a the equivalent of a
berserk mode where they no longer care about their health.

Graphically:

             no longer cares about health
           and will fight to it's own demise -,
        ,-------------------------------------'
        v
      y                 normal combat -,
    1 | ;          ; <-----------------'
      | '          ;
      |  '       ,'   discouraged as health is low -,
      |  '     ,'  <--------------------------------'
      |   ',-'
    0 |______________ x
       0            1